+++ CHANGELOG.md
@@ -20,6 +20,7 @@ Enhancements
 * `CustomType` derive macro now handles `Option` fields (thanks [`@agersant`](https://github.com/agersant) [#1011](https://github.com/rhaiscript/rhai/pull/1011)).
 * `Engine::eval_binary_op` is added to quickly compare two `Dynamic` values.
 * Better handling for 32-bit architectures and enhanced safety by replacing casts with `try_from` (thanks [`@therealprof`](https://github.com/therealprof) [#1009](https://github.com/rhaiscript/rhai/pull/1009)).
+* `Engine::register_custom_syntax_without_look_ahead_raw` is added to allow the use of `$raw$` in custom synax, which returns the script text character-by-character without processing, by-passing the tokenizer. This in turn allows for parsing arbitrary syntax.
 
 
 Version 1.22.2